git-svn-id: http://62.90.102.41/svn/NiceHouse/trunk@194 1431c75d-6ac1-4f6d-bc36-6b0d284ea005

+++ Management/models.py
@@ -1459,7 +1459,7 @@ class Sale(models.Model):
     clients_phone = models.CharField(ugettext('phone'), max_length = 10)
     pc_base = models.FloatField(editable=False, null=True)
     pb_dsp = models.FloatField(editable=False, null=True)
-    zdb = models.FloatField(editable=False, null=True)#zilber discount bonus
+    zdb = models.IntegerField(editable=False, null=True)#zilber discount bonus
     c_final = models.FloatField(editable=False, null=True)
     price_final = models.IntegerField(editable=False, null=True)
     employee_pay = models.DateField(ugettext('employee_paid'), editable=False)
